+++ src/front/js/layout.js
@@ -12,9 +12,10 @@ import { Footer } from "./component/footer";
 import { Europa } from "./pages/europe";
 import { Asia } from "./pages/asia";
 import { Africa } from "./pages/africa";
-import { Registro} from "./component/formularioRegistro";
-import {LoginVista} from "./pages/loginvista";
-import { PerfilVista } from "./pages/vistaPerfil";
+import { Registro} from "./component/formularioRegistro.js";
+import {LoginVista} from "./component/loginvista.js";
+import { PerfilVista } from "./pages/vistaPerfil.js";
+import { RecuperacionContraseña  } from "./component/recuperacionContraseña.js";
 
 //create your first component
 const Layout = () => {
@@ -36,7 +37,8 @@ const Layout = () => {
                        <Route element={<Africa />} path="/africa" />
                        <Route element = {<LoginVista />} path= "/login" />
                        <Route element = {<Registro/>} path= "/Registro" />
-                       <Route element = {<PerfilVista/>} path= "/vistaPerfil" />
+                       <Route element = {<PerfilVista/>} path= "/Perfil" />
+                       <Route element = {<RecuperacionContraseña/>} path= "/recuperacion-contrasena" />
 
                         <Route element={<h1>Not found!</h1>} />
                     </Routes>
